requirements.txt
langchain
langchain-community
langchain-openai
langchain-groq
pypdf
faiss-cpu
python-dotenv
streamlit
langchain_community
langserve 
fastapi
uvicorn 
sse_starlette
bs4
pypdf
chromadb
faiss-cpu
Wikipedia
arxiv
langchainhub
langchain-classic 
groq
Beautifulsoup4
langchain-ollama
sentence_transformers 
